Silence vim E185 ("Cannot find color scheme")
It is difficult to detect if the color scheme plugins has been properly
installed and loaded. On initial load up, the Dracula color scheme will
not be installed and loaded until after `PlugInstall`. Silence E185 such
that the "Press ENTER or type command to continue" prompt does not
disrupt vim load up.